refactored for renamed inventory file
