+++ .gitignore
@@ -100,3 +100,7 @@ ENV/
 
 # mypy
 .mypy_cache/
+
+# inventory data
+data/
+log/
