+++ roocs_utils/etc/roocs.ini
@@ -1,5 +1,5 @@
 [config_data_types]
-lists = facet_rule known_projects
+lists = facet_rule
 dicts = mappings attr_defaults
 ints = n_per_batch
 floats =
@@ -37,6 +37,7 @@ facet_rule = mip_era activity_id institution_id source_id experiment_id member_i
 mappings =
     variable:variable_id
 
+
 [project:cordex]
 base_dir = /badc/cordex/data
 file_name_template = {__derive__var_id}_{CORDEX_domain}_{driving_model_id}_{experiment}_{driving_model_ensemble_member}_{model_id}_{rcm_version_id}_{frequency}_{__derive__time_range}.{__derive__extension}
@@ -50,6 +51,7 @@ attr_defaults =
     frequency:no-freq
 facet_rule = project product domain institute driving_model experiment ensemble rcm_name rcm_version time_frequency variable
 mappings =
+
 
 [project:c3s-cmip5]
 project_name = c3s-cmip5
@@ -71,6 +73,7 @@ error_pickle =  %(inventory_dir)s/%(project_name)s-errors.pickle
 inventory_pickle =  %(inventory_dir)s/%(project_name)s-inventory.pickle
 local_inventory_file = %(inventory_dir)s/%(project_name)s-inventory-local.yml
 c3s_inventory_file = %(inventory_dir)s/%(project_name)s-inventory-c3s.yml
+
 
 [project:c3s-cmip6]
 project_name = c3s-cmip6
